add d.ts to hash-name, move utlities to components
