+++ packages/react/scripts/rollup/hash-name.d.ts
@@ -1,0 +1,5 @@
+export declare function hashCode(input: string): string;
+export declare function generateScopedName(
+  name: string,
+  filename: string,
+): string;
