+++ packages/react/scripts/rollup/hash-name.js
