+++ packages/react/src/components/legacy/LegacyTextArea/TextArea.tsx
@@ -1,8 +1,11 @@
 import React, { forwardRef, useId, useState } from 'react';
 import type { ChangeEvent, ForwardedRef, TextareaHTMLAttributes } from 'react';
 
-import { InputWrapper } from '../../_InputWrapper';
-import type { ReadOnlyVariant_, CharacterLimit } from '../../_InputWrapper';
+import { InputWrapper } from '../../utilities/InputWrapper';
+import type {
+  ReadOnlyVariant_,
+  CharacterLimit,
+} from '../../utilities/InputWrapper';
 
 import classes from './TextArea.module.css';
