+++ packages/react/src/components/legacy/LegacyTextField/TextField.tsx
@@ -10,8 +10,11 @@ import type {
 import { NumericFormat, PatternFormat } from 'react-number-format';
 
 import { isNumericFormat, isPatternFormat } from '../../../utils';
-import { InputWrapper } from '../../_InputWrapper';
-import type { ReadOnlyVariant_, CharacterLimit } from '../../_InputWrapper';
+import { InputWrapper } from '../../utilities/InputWrapper';
+import type {
+  ReadOnlyVariant_,
+  CharacterLimit,
+} from '../../utilities/InputWrapper';
 
 export type LegacyTextFieldProps = {
   /**
