+++ packages/react/src/components/utilities/InputWrapper/Icon.module.css
