+++ packages/react/src/components/utilities/InputWrapper/InputWrapper.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from 'react';
 import cn from 'classnames';
 
 import utilityClasses from '../../utils/utility.module.css';
-import { ErrorMessage, Paragraph } from '../Typography';
+import { ErrorMessage, Paragraph } from '../../Typography';
 
 import type { ReadOnlyVariant_, InputVariant_ } from './utils';
 import { getVariant } from './utils';
