+++ packages/react/src/components/utilities/InputWrapper/SearchIcon.tsx
